+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0
 
 - Make sure to run after `.dart.js` files are generated.
+- Add `debug` option to print out all files that are excluded and why.
 
 ## 1.0.1